adding tests

+++ app/app/graphql/mutations/delete_comment.rb
@@ -5,13 +5,19 @@ class Mutations::DeleteComment < Mutations::BaseMutation
   field :message, String, null: false
 
   def resolve(id:)
-    comment = Comment.find(id)
-    post_id = comment.post.id
-    comment.destroy
-    CableHelpers::Comments.broadcast_comments(post_id)
-    return {
-      message: 'successfully deleted comment'
-    }
+    if is_authorized?
+      comment = Comment.find(id)
+      post_id = comment.post.id
+      comment.destroy
+      CableHelpers::Comments.broadcast_comments(post_id)
+      return {
+        message: 'successfully deleted comment'
+      }
+    else
+      return {
+        errors: ['not authorized']
+      }
+    end
   end
 
 end
